+++ src/game/tests.rs
@@ -508,6 +508,7 @@ mod attachment_targets;
 mod attack_deterrents;
 mod attack_restrictions;
 mod attacked_last_turn;
+mod attacking_walls;
 mod aura_death;
 mod aura_tap_triggers;
 mod aura_upkeep;